PR changes Move PlatformPill to page/common Clarify MapPage doc Format HazardMarkerLayer doc Make settings showConfirmationDialog async
