+++ lib/page/common/platform_pill.dart
